Better support for dual RDBE systems
The /usr2/control/rdbead.ctl holds the FS addresses for RDBE0, RDBE1
so it's only natural that vlbis supports:

    > fs rdbe 0

to address whatever RDBE0 is currently set to.
This code supports any number of RDBEs, not limited to 0 or 1 ...

Also, when talking to rdbe_server, the terminating ';' is essential or else
it'll give a syntax error. So it's only natural that vlbish adds the
semicolon for you in case you forgot ...