Fixup parser a bit
During usage ran into a nagging thing:

    > alias a localhost:2620,sockbuf=320M,n=4,proto=udpsnor

would execute fine, but then _using_ it:

    > resolve a

would result in a parse error:

    localhost:2620,sockbuf=320M,n=4,proto=udpsnort
                              ^
    0 tokens matched here

Tracing the tokenizer, turns out that after
    sockbuf=
there are two matches:
    320  (token type int, value 320)
    320M (token type text, value '320M')

and the code then chooses the first.
Changed the tokenizer to sort the matches by size descending such that
if there are longer matches, that will be the token. If both matches are
the same length, the order is stable. The matching token will be the one
appearing first in the token list.

Added command line option to control printing tokenizer state/progress.

Whilst at it, made a few statements simpler:
    pos += (mo.end() - mo.start())
simplified to:
    pos = mo.end()

and the implementation function for printing queries, if so commanded,
is now done by reusing the print_debug() debug print function generating
function.